requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.21.0
pyyaml>=6.0.0

# Google Cloud dependencies
google-cloud-storage>=2.10.0
google-cloud-firestore>=2.11.0
google-cloud-secret-manager>=2.16.0
google-cloud-aiplatform>=1.25.0
google-cloud-functions>=1.19.0
google-cloud-run>=0.5.0
google-cloud-logging>=3.8.0
functions-framework>=3.4.0
google-api-python-client>=2.100.0
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0

# Machine Learning dependencies
scikit-learn>=1.3.0
nltk>=3.8.0

# HTTP and Authentication
requests>=2.31.0
requests-ntlm>=1.2.0

# Development dependencies
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0

# Utilities
python-dotenv>=1.0.0
metaphone>=0.6  # For phonetic matching
dataclasses>=0.6  # For Python 3.6 compatibility if needed

# Google Cloud IAM dependencies
google-cloud-iam>=2.12.0  # For IAM and service account management

# Google Cloud Pub/Sub dependencies
google-cloud-pubsub>=2.18.0

# Google Cloud Scheduler dependencies
google-cloud-scheduler>=2.12.0

# Backoff dependencies
backoff==2.2.1

# TZData dependencies
tzdata==2023.3
